Revamp API to use encodings and return a stream
The API is now reworked to detect and return a successful response's
mimetype and response parameters, and the charset is detected to
return an appropriate flexi-stream.

Gemini-request's errors are also reworked to only return an error in
likely cases where the server is misbehaving -- mainly, inappropriate
response headers, too many redirects, and bad encoding.

Further, the API now has three main functions instead of two --
the primitive, gemini-request-stream*, and then the more higher-level
gemini-request-stream and gemini-request. The first two always returns
a stream as the primary value while the second function returns the
response body as a string or octet simple-vector.